权限中心 删除字典分类

+++ toucan-shopping-modules/adminAuth/toucan-shopping-admin-auth-service-dao/src/main/java/com/toucan/shopping/modules/admin/auth/mapper/DictCategoryMapper.java
@@ -59,7 +59,7 @@ public interface DictCategoryMapper {
      * @param id
      * @return
      */
-    int deleteById(Long id);
+    int deleteById(Integer id);
 
     /**
      * 根据应用查询列表
