+++ toucan-shopping-modules/adminAuth/toucan-shopping-admin-auth-service-dao/src/main/java/com/toucan/shopping/modules/admin/auth/service/DictCategoryService.java
@@ -57,7 +57,7 @@ public interface DictCategoryService {
      * @param id
      * @return
      */
-    int deleteById(Long id);
+    int deleteById(Integer id);
 
 
     /**
