+++ toucan-shopping-modules/adminAuth/toucan-shopping-admin-auth-service-dao/src/main/java/com/toucan/shopping/modules/admin/auth/service/impl/DictCategoryServiceImpl.java
@@ -72,7 +72,7 @@ public class DictCategoryServiceImpl implements DictCategoryService {
     }
 
     @Override
-    public int deleteById(Long id) {
+    public int deleteById(Integer id) {
         return dictCategoryMapper.deleteById(id);
     }
